# docs/swagger.yaml
basePath: /
definitions:
  requests.CreateCharacterRequest:
    properties:
      class_id:
        type: integer
      level:
        maximum: 20
        minimum: 1
        type: integer
      name:
        maxLength: 200
        type: string
      pronouns:
        maxLength: 32
        type: string
      race_id:
        type: integer
    required:
    - class_id
    - level
    - name
    - race_id
    type: object
...
  requests.UpdateCharacterRequest:
    properties:
      class_id:
        type: integer
      level:
        maximum: 20
        minimum: 1
        type: integer
      name:
        maxLength: 200
        type: string
      pronouns:
        maxLength: 32
        type: string
      race_id:
        type: integer
    required:
    - class_id
    - level
    - name
    - race_id
    type: object
...
  responses.CharacterPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.CharacterResponse'
        type: array
      meta:
        $ref: '#/definitions/responses.ResponseMeta'
    type: object
  responses.CharacterResponse:
    properties:
      class:
        $ref: '#/definitions/responses.ClassResponse'
      class_id:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      profile_picture:
        $ref: '#/definitions/responses.FileResponse'
      pronouns:
        type: string
      race:
        $ref: '#/definitions/responses.RaceResponse'
      race_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  responses.ClassResponse:
    properties:
      hit_point_die_value:
        type: integer
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      primary_ability:
        type: string
      saves:
        type: string
      short_description:
        type: string
      subclasses:
        items:
          $ref: '#/definitions/responses.SubclassResponse'
        type: array
    type: object
  responses.Data:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  responses.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  responses.FileResponse:
    properties:
      created_at:
        type: string
      enterprise_uid:
        type: string
      file_location:
        type: string
      filename:
        type: string
      id:
        example: 1
        type: integer
      model:
        type: string
      model_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  responses.RaceResponse:
    properties:
      base_speed:
        type: integer
      creature_type:
        type: string
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      short_description:
        type: string
      size:
        type: string
    type: object
...
  responses.ResponseMeta:
    properties:
      page:
        example: 3
        type: integer
      page_size:
        example: 25
        type: integer
      total_count:
        example: 300
        type: integer
    type: object
...
  responses.SubclassResponse:
    properties:
      class_id:
        type: integer
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      short_description:
        type: string
    type: object
info:
  contact:
    email: jbrokenshire0306@gmail.com
    name: Jared Brokenshire
  description: A management solution for Dungeons & Dragons character sheet information
  title: D&D Character Dashboard
  version: 0.0.1
paths:
...
  /characters:
    get:
      consumes:
      - application/json
      description: List characters (paginated)
      operationId: characters-list
      parameters:
      - description: Search characters by name
        in: query
        name: search
        type: string
      - description: The page number
        in: query
        name: page
        type: integer
      - description: The numbers of items to return. Max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CharacterPaginatedResponse'
      summary: List characters
      tags:
      - Character Actions
    post:
      consumes:
      - application/json
      description: Create character
      operationId: characters-create
      parameters:
      - description: Character information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreateCharacterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CharacterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create character
      tags:
      - Character Actions
  /characters/{id}:
    delete:
      consumes:
      - application/json
      description: Delete character
      operationId: characters-delete
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Delete character
      tags:
      - Character Actions
    get:
      consumes:
      - application/json
      description: Get character by ID
      operationId: characters-get
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CharacterResponse'
      summary: Get character by ID
      tags:
      - Character Actions
    put:
      consumes:
      - application/json
      description: Update character
      operationId: characters-update
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Character information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateCharacterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CharacterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update character
      tags:
      - Character Actions
  /characters/{id}/upload/profile-picture:
    post:
      consumes:
      - application/json
      description: Upload character profile picture
      operationId: characters-upload-profile-picture
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Upload character profile picture
      tags:
      - Character File Actions
...
swagger: "2.0"